requirements.txt
Flask
Werkzeug
boto3
Flask-Cors
python-dotenv